requirements.txt
django>=5.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-anymail[sendgrid]
django-storages[boto3]
django-celery-beat
django-filter
django-extensions
celery[redis]
redis
anthropic
stripe
pyfcm
Pillow
qrcode[pil]
whitenoise
gunicorn
psycopg2-binary
dj-database-url
python-decouple
boto3
requests
sentry-sdk[django]
